# Optimize the starts_with filter