# Pull Request

Previously, the `STARTS_WITH` filter performed a deep prefix scan on the facet tree, starting from the leaf level (level 0). While this approach works, it’s inefficient in cases where a higher-level node already matches the prefix and contains all relevant document IDs. In such cases, scanning each leaf node individually is unnecessary and costly.

## Related issue

Fixes https://github.com/meilisearch/meilisearch/issues/5801

## What does this PR do?

This PR introduces the use of a recursive algorithm to optimize the `STARTS_WITH` filter by converting it into a range search. Specifically, instead of performing a deep scan, the filter now searches for facets within a prefix-based range. For example, querying facets starting with `baba` will now look for facets in the range `baba..babb`. This doesn't change anything for users.

This is a simple yet effective optimization that significantly reduces scan time in cases where the prefix is common across many entries.

## Benchmarks

The new approach offers substantial speed improvements for short or common prefixes. As expected, the previous implementation remains faster for unique prefixes, but the new algorithm still performs very well in those cases and offers a considerable net gain overall.

 Prefix length | Original | New | Speedup
-- | -- | -- | --
1 letter | 1168.2 | 263.84 | 4.4x (faster)
2 letters | 198.68 | 49.542 | 4.0x (faster)
3 letters | 136.64 | 40.057 | 3.4x (faster)
6 letters | 23.567 | 27.907 | 0.8x (slower)

## Design Considerations

While it’s slightly slower for highly specific prefixes (e.g., 6+ characters), the performance hit is minimal. I considered enabling the new logic only for shorter prefixes (e.g., ≤6 characters), but this could lead to confusing behavior and inconsistent performance. Especially in datasets with uneven value distributions, where adding a single character could drastically change query speed.

Ultimately, I believe the simplicity and overall performance gains of the new approach outweigh the downsides, and it offers a more robust experience for the majority of use cases.