@Mubelotix 2025-08-05
You were right, and it turned out `find_docids_of_facet_within_bounds` already accepted Bytes.